requirements.txt
streamlit
opencv-python-headless
Pillow
torch
transformers
torchvision
